file/file.go
package file

import (
	"io/ioutil"
	"log"
	"os"
	"path/filepath"
	"strings"

	"github.com/CatchZeng/gutils/convert"
)

// Exists reports whether the file or directory exists.
func Exists(path string) (existed bool) {
	existed, _ = Exist(path)
	return
}

// Exist reports whether the file or directory exists.
func Exist(path string) (existed bool, isDir bool) {
	info, err := os.Stat(path)
	if err != nil {
		return !os.IsNotExist(err), false
	}
	return true, info.IsDir()
}

// Mode get file mode
func Mode(path string) os.FileMode {
	fileInfo, err := os.Stat(path)
	if err != nil {
		return 0755
	}
	return fileInfo.Mode()
}

// WriteStringToFile write string to file
func WriteStringToFile(content, path string, mode os.FileMode) (err error) {
	bytes := convert.StringToBytes(content)
	return ioutil.WriteFile(path, bytes, mode)
}

// AppendStringToFile append string to file
func AppendStringToFile(content string, filePath string) error {
	f, err := os.OpenFile(filePath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
	if err != nil {
		return err
	}
	defer f.Close()
	if _, err := f.WriteString(content); err != nil {
		return err
	}
	return nil
}

// GetDirList get directory list
func GetDirList(path string) ([]string, error) {
	return GetDirListWithFilter(path, nil)
}

// Filter file filter
type Filter func(os.FileInfo) bool

// GetDirListWithFilter get directory list with filter
func GetDirListWithFilter(path string, filter Filter) ([]string, error) {
	var dirList []string

	paths, err := filepath.Glob(filepath.Join(path, "*"))

	log.Printf("paths: %v", paths)

	for _, value := range paths {
		f, err := os.Stat(value)
		if err != nil {
			return dirList, err
		}
		if filter != nil && !filter(f) {
			continue
		}
		if f.IsDir() {
			dir := strings.Replace(value, path, "", 1)
			if strings.HasPrefix(dir, "/") {
				dir = strings.Replace(dir, "/", "", 1)
			}
			dirList = append(dirList, dir)
		}
	}

	return dirList, err
}
